backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Senator",email = "144132593+itsSenat0r@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi[all] (>=0.120.1,<0.121.0)",
    "motor (>=3.7.1,<4.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose[crypto] (>=3.5.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "torch (>=2.9.1,<3.0.0)",
    "silero-stress (>=1.1,<2.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "sounddevice (>=0.5.3,<0.6.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "websocket-client (>=1.9.0,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
